Refactor SimulationHandler to separate class
